refactor: ajustando componente
